requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20230802.0=h5da7b33_2
absl-py=2.1.0=py38haa95532_0
aiohttp=3.9.5=py38h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
async-timeout=4.0.3=py38haa95532_0
attrs=23.1.0=py38haa95532_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
blinker=1.6.2=py38haa95532_0
bottleneck=1.3.7=py38h9128911_0
brotli=1.0.9=h2bbff1b_8
brotli-bin=1.0.9=h2bbff1b_8
brotli-python=1.0.9=py38hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2024.3.11=haa95532_0
cachetools=5.3.3=py38haa95532_0
certifi=2024.6.2=py38haa95532_0
cffi=1.16.0=py38h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py38haa95532_0
colorama=0.4.6=py38haa95532_0
cryptography=42.0.5=py38h89fc84f_1
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
eigen=3.4.0=h59b6b97_0
ffmpeg=4.2.2=he774522_0
filelock=3.13.1=py38haa95532_0
fonttools=4.51.0=py38h2bbff1b_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.0=py38h2bbff1b_0
fsspec=2024.3.1=py38haa95532_0
gdown=5.1.0=pypi_0
giflib=5.2.1=h8cc25b3_3
glib=2.78.4=hd77b12b_0
glib-tools=2.78.4=hd77b12b_0
google-auth=2.29.0=py38haa95532_0
google-auth-oauthlib=0.5.2=py38haa95532_0
grpc-cpp=1.48.2=h6772dbd_4
grpcio=1.48.2=py38h6772dbd_4
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
gtest=1.14.0=h59b6b97_1
hdf5=1.12.1=h51c971a_3
huggingface_hub=0.23.1=py38haa95532_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.7=py38haa95532_0
ignite=0.5.0.post2=py_0
importlib-metadata=7.0.1=py38haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
joblib=1.4.2=py38haa95532_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=py38hd77b12b_0
krb5=1.20.1=h5b6d351_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libglib=2.78.4=ha17d25a_0
libiconv=1.16=h2bbff1b_3
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libprotobuf=3.20.3=h23ce68f_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=he774522_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_2
libxslt=1.1.37=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_1
markdown=3.4.1=py38haa95532_0
markupsafe=2.1.3=py38h2bbff1b_0
matplotlib=3.5.3=py38haa244fe_2
matplotlib-base=3.5.3=py38hd77b12b_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.8=py38h2bbff1b_0
mkl_random=1.2.4=py38h59b6b97_0
multidict=6.0.4=py38h2bbff1b_0
numexpr=2.8.4=py38h7b80656_1
numpy=1.24.3=py38h79a8e48_1
numpy-base=1.24.3=py38h8a87ada_1
oauthlib=3.2.2=py38haa95532_0
opencv=4.6.0=py38ha36de5b_5
opencv-contrib-python=4.9.0.80=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.14=h827c3e9_0
packaging=23.2=py38haa95532_0
pandas=1.5.3=py38hf11a4ad_0
pcre2=10.42=h0ff8eda_1
pillow=10.3.0=py38h2bbff1b_0
pip=24.0=py38haa95532_0
platformdirs=3.10.0=py38haa95532_0
ply=3.11=py38_0
pooch=1.7.0=py38haa95532_0
protobuf=3.20.3=py38hd77b12b_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.8.0=py38haa95532_0
pyopenssl=24.0.0=py38haa95532_0
pyparsing=3.0.9=py38haa95532_0
pyqt=5.15.10=py38hd77b12b_0
pyqt5-sip=12.13.0=py38h2bbff1b_0
pysocks=1.7.1=py38haa95532_0
python=3.8.18=h1aa4202_0
python-dateutil=2.9.0post0=py38haa95532_2
python_abi=3.8=2_cp38
pytorch=1.11.0=py3.8_cuda11.3_cudnn8_0
pytorch-fid-wrapper=0.0.4=pypi_0
pytorch-model-summary=0.1.1=py_0
pytorch-mutex=1.0=cuda
pytz=2024.1=py38haa95532_0
pyyaml=6.0.1=py38h2bbff1b_0
qt-main=5.15.2=h879a1e9_9
qt-webengine=5.15.9=h5bd16bc_7
qtwebkit=5.212=h2bbfb41_5
re2=2022.04.01=hd77b12b_0
requests=2.32.2=py38haa95532_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
safetensors=0.4.2=py38h0010962_0
scikit-learn=1.3.0=py38h4ed8f06_0
scipy=1.10.1=py38hdcfc7df_1
seaborn=0.12.2=py38haa95532_0
setuptools=69.5.1=py38haa95532_0
sip=6.7.12=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.5=pypi_0
sqlite=3.45.3=h2bbff1b_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.15.1=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py38haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
timm=0.9.10=pyhd8ed1ab_0
tomli=2.0.1=py38haa95532_0
torch-tb-profiler=0.4.3=pypi_0
torchaudio=0.11.0=py38_cu113
torchvision=0.12.0=py38_cu113
tornado=6.4.1=py38h827c3e9_0
tqdm=4.66.4=py38h9909e9c_0
ttach=0.0.3=pyhd8ed1ab_0
typing-extensions=4.11.0=py38haa95532_0
typing_extensions=4.11.0=py38haa95532_0
unicodedata2=15.1.0=py38h2bbff1b_0
urllib3=2.2.2=py38haa95532_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
werkzeug=3.0.3=py38haa95532_0
wheel=0.43.0=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.9.3=py38h2bbff1b_0
zipp=3.17.0=py38haa95532_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
